backend/MTSSummarizerBackend/requirements.txt
adrf==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-property==0.2.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
fake-useragent==2.2.0
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
multidict==6.4.3
oauthlib==3.2.2
packaging==25.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
uvicorn-worker==0.3.0
yarl==1.20.0
